requirements.txt
python-dotenv
llmwhisperer-client